Disparo de alerta + reportes de alertas

+++ src/main/java/co/edu/ucc/sipnat/vista/ModuloProyecto/MbVerProyecto.java
@@ -189,7 +189,7 @@ public class MbVerProyecto implements Serializable {
         series1.setLabel("Codigo: " + id);
         for (DatosSensor dse : dses) {
             SimpleDateFormat formatter = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-            series1.set(formatter.format(dse.getFechaRecoleccion()), Integer.parseInt(dse.getDato()));
+            series1.set(formatter.format(dse.getFechaRecoleccion()), Double.parseDouble(dse.getDato()));
         }
         model.addSeries(series1);
         return model;
